Plots with CairoMakie

Before calling one of these functions you must load CairoMakie.jl with

using CairoMakie

Plot series of endodormancy break and budburst Dates

Phenology.Plot_Pheno_Dates_EB_BB — Method

Phenology.Plot_Pheno_Dates_EB_BB(date_vecEB::Vector{Date},
    date_vecBB::Vector{Date},
    CPO;
    sample_EB=nothing,
    sample_BB=nothing,
    station_name="",
    YearCut=nothing,
    save_file=nothing,
    loaded_data=nothing,
    display_legend=true)

Plot the dormancy break and budburst dates from the series date_vecEB and date_vecBB and if precised from the samples sample_EB and sample_BB.

Phenology.Plot_Pheno_Dates_EB_BB — Method

Phenology.Plot_Pheno_Dates_EB_BB(date_vecsEB, date_vecsBB, CPO;
    EB_colors=nothing,
    BB_colors=nothing,
    EB_label=nothing,
    BB_label=nothing,
    dashindexes=Integer[],
    size=(800, 400),
    breakpoints=true,
    comments="")

Plot the dormancy break and budburst dates from the series date_vecsEB and date_vecsBB.

Phenology.PlotHistogram — Function

Phenology.PlotHistogram(date_vecEB::Vector{Date},
    date_vecBB::Vector{Date},
    CPO,
    year;
    sample_EB=nothing,
    sample_BB=nothing,
    station_name="",
    LineHeight=0.15,
    stationlegend=false,
    horline=true,
    PlotQuantile=false)

For the year precised, plot the histogram of the endodormancy break and budburst dates from the samples sample_EB and sample_BB. It plots also the phenological dates of that year from date_vecEB and date_vecBB as vertical lines.

Plot Freezing Risk

Phenology.Plot_Freeze_Risk — Function

Phenology.Plot_Freeze_Risk(TN_vecs, dates_vecs_TN, date_vecsBB;
    CPO=(8, 1),
    colors=nothing,
    label=nothing,
    size=(800, 400),
    threshold=-2)

For multiple daily minimal temperatures series TN_vecs and their respective dates dates_vecs_TN and BudBurst dates date_vecsBB, plot the number of days with TN < threshold per year.

Phenology.Plot_Freeze_Risk_sample — Function

Phenology.Plot_Freeze_Risk_sample(TN_vecs,
    Date_vec,
    date_vecsBB;
    threshold=-2.,
    PeriodOfInterest=Month(3),
    CPO=(10, 30),
    size_fig=(1300, 400))

For a sample of daily minimal temperatures series TN_vecs and budburst series date_vecsBB with their date Date_vec plot the yearly mean, variance and maximum of number of days with TN < threshold.

Phenology.Plot_Freeze_Risk_Bar — Function

Phenology.Plot_Freeze_Risk_Bar(TN_vec, dates_vec_TN, date_vecBB;
    CPO=(8, 1),
    color=nothing,
    label=nothing,
    size=(800, 400),
    threshold=-2)

For a series of daily minimal temperatures TN_vec and his respective dates dates_vec_TN and BudBurst dates date_vecsBB, plot the number of days with TN < threshold per year.