Freezing risk

Phenology.FreezingRisk — Function

FreezingRisk(TN_vec::AbstractVector, Date_vec::AbstractVector{Date}, BB::Date; threshold=-2., PeriodOfInterest=Month(3), CPO=(10, 30))
FreezingRisk(df::DataFrame, BB; threshold=-2., PeriodOfInterest=Month(3), CPO=(8, 1))

For a given budburst date BB return the number of days with minimal temperature inferior to threshold for the period PeriodOfInterest after this date but always before the next chilling period onset CPO (next phenological cycle).

Phenology.FreezingRiskMatrix — Method

FreezingRiskMatrix(TN_vec, Date_vec, date_vecBB::AbstractVector{Date}; threshold=-2., PeriodOfInterest=Month(3), CPO=(10, 30))

For a given budburst date vector date_vecBB return the number of days with minimal temperature inferior to threshold for each year in a N x 2 matrix

Phenology.FreezingRiskMatrix — Method

FreezingRiskMatrix(TN_vecs, Date_vec, date_vecsBB; threshold=-2., PeriodOfInterest=Month(3), CPO=(10, 30))

For a sample of daily minimal temperatures series TN_vecs and their respectives budburst dates date_vecsBB return a matrix with the number of scenarios which have n days with minimal temperature inferior to threshold for n in row and the year in column.