Phenological models for each species

Apple

Phenology.AppleModel — Type

AppleModel(CPO::Tuple{Integer,Integer}=(10, 30),
    chilling_model::AbstractAction=TriangularAction(1.1, 20.),
    chilling_target::AbstractFloat=56.0,
    forcing_model::AbstractAction=ExponentialAction(9.0),
    forcing_target::AbstractFloat=83.58)

Structure which contains the parameters for a phenelogical apple model. The default values are the ones suggested in (Legave et al., 2013) (F1 Gold 1).

Phenology.Apple_Phenology_Pred — Function

Apple_Phenology_Pred(TG_vec::AbstractVector, date_vec::AbstractVector{Date}; model::AppleModel=AppleModel())
Apple_Phenology_Pred(df::DataFrame; model::AppleModel=AppleModel())
Apple_Phenology_Pred(file_TG::String; model::AppleModel=AppleModel())

From a series of daily average temperature TG_vec, its dates in date_vec and an apple phenology model model, return the endodormancy break dates and budburst dates in two vectors respectively. The temperatures and dates data can be included in a dataframe (second method) or in a .txt file (third method). See Temperatures data compatibility for further explanation about the way to input temperatures data.

Grapevine (BRIN)

Phenology.BRIN_Model — Type

BRIN_Model(CPO::Tuple{Integer,Integer}=(8, 1),
    Q10::AbstractFloat=2.17,
    chilling_target::AbstractFloat=119.0,
    T0Bc::AbstractFloat=8.19,
    TMBc::AbstractFloat=25.,
    forcing_target::AbstractFloat=13236)
BRIN_Model(date_vec, x, date_vecBB::AbstractVector{Date}; p0=[100., 8000.])
BRIN_Model(date_vec, x, years, doy; p0=[100., 8000.])
BRIN_Model(x_vec::AbstractVector, n_train::AbstractVector; p0=[100., 8000.])

Structure which contains the parameters for a phenelogical BRIN model for grapevine. The default values are the ones suggested in (García de Cortázar-Atauri et al., 2009) (table 5 p 323). With the three last methods you can initiate a BRIN_Model by fitting on temperatures, dates et bud burst dates data (see Training BRIN models). It is done with a differential evolution optimization algorithm.

Phenology.Vine_Phenology_Pred — Function

Vine_Phenology_Pred(TN_vec::AbstractVector, TX_vec::AbstractVector, date_vec::AbstractVector{Date}; model::BRIN_Model=BRIN_Model())
Vine_Phenology_Pred(file_TN::String, file_TX::String; model::BRIN_Model=BRIN_Model())
Vine_Phenology_Pred(df_TN::DataFrame, df_TX::DataFrame; model::BRIN_Model=BRIN_Model())
Vine_Phenology_Pred(df::DataFrame; model::BRIN_Model=BRIN_Model())

From a series of TN TN_vec, a series of TX TX_vec, their dates in date_vec and a vine phenology model model, return the endodormancy break dates and budburst dates in two vectors respectively. The temperatures and dates data can be included in two .txt file, two different dataframes or one dataframe with the two type of temperature. See Temperatures data compatibility for further explanation about the way to input temperatures data.